msm: sps: Add IPC logging for SPS driver
Add the IPC logging for SPS driver, which helps monitoring the
interactions between BAM clients and SPS driver and helps
tracking the BAM status.

Change-Id: Ic713acf8052da2725b5d0d870073dc0787acd0fb
Signed-off-by: Yan He <yanhe@codeaurora.org>

msm: sps: add the checking of ZLT in case of an empty pipe

Check if a ZLT descriptor is the last pending descriptor in
the desc FIFO when determine if the desc FIFO of that pipe is
empty.

Change-Id: Ib8174953d6ad7d886f47f21e243e985b79ef41fb
Signed-off-by: Yan He <yanhe@codeaurora.org>

msm: sps: support ZLT injection

Add the support to inject a ZLT descriptor with EOT, so that a peripheral
of BAM can respond to BAM to close pending descriptors.

Change-Id: I0db774de8f00cca02e895f45530a84f47c9dfe00
Signed-off-by: Yan He <yanhe@codeaurora.org>

msm: sps: Add loglevel property in IPC logging for SPS driver

IPC logging for SPS driver, helps monitoring the interactions
between BAM clients and SPS driver and helps tracking the BAM
status. Add a loglevel field such that clients can specify the
loglevel to be captured in IPC logging.

Change-Id: Ib3af72485caa48940efca74b2cec2464a09ffed8
Signed-off-by: Rama Krishna Phani A <rphani@codeaurora.org>

msm: sps: avoid register read for ignored debugging messages

When the detailed logging is not enabled, pr_debug() will not
output the logging message. However, the parameters of pr_debug()
are still calculated. Thus, the parameters involving register read
could consume significant CPU time when the function with this kind
of logging message is invoked frequently. The fix here is to ignore
the operation of logging involving register access when the
detailed logging is disabled.

Change-Id: I1aff13c6f054c319034d7a47356bdd345741f159
Signed-off-by: Yan He <yanhe@codeaurora.org>

msm: sps: add the checking of ZLT in case of an empty pipe

Check if a ZLT descriptor is the last pending descriptor in
the desc FIFO when determine if the desc FIFO of that pipe is
empty.

Change-Id: Ib8174953d6ad7d886f47f21e243e985b79ef41fb
Signed-off-by: Yan He <yanhe@codeaurora.org>

Conflicts:
	drivers/platform/msm/sps/sps.c
	include/linux/msm-sps.h